requirements.txt
discord.py
python-dotenv  
pymongo  
aiohttp
PyPDF2
subprocess